+++ src/h3sed/metadata.py
@@ -7,7 +7,7 @@ This file is part of h3sed - Heroes3 Savegame Editor.
 Released under the MIT License.
 
 @created     22.03.2020
-@modified    14.09.2024
+@modified    04.12.2024
 ------------------------------------------------------------------------------
 """
 from collections import defaultdict, OrderedDict