+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---------------
 - fix saving slots that were initially empty then populated then emptied again
 - fix tracking changes over undo-redo after saving
+- fix error after confirming to save changes when closing file
 
 
 2.7, 2024-09-14